requirements.txt
langchain==0.1.12
langchain-groq==0.0.6
streamlit==1.32.2
python-dotenv==1.0.1 